Add handling for keyword fields with a maximum length of 1024 characters
mb_substr is used to perform substring extraction; ElasticSearch uses UTf-8 exclusively, so this is the encoding given to mb_substr. In addition, in order to check string length, strlen is first used (it is constant time), and only if strlen exceeds the character limit is mb_substr used (which is linear time).

The keyword handing is added to most of the locations it is found in the Python agent (https://github.com/elastic/apm-agent-python/search?q=keyword_field&unscoped_q=keyword_field); a few are omitted where it doesn't seem to make any reasonable sense (e.g. the PHP version, and exceptions codes; while both technically can exceed the character limit, they should really only do so in extremely atypical situations).

The keyword handling was also omitted from tag names (simply because I wasn't 100% confident in modifying that code).